python/transferAndRemoteRun/requirements.txt
paramiko
fabric
telnetlib3 
pywinrm
